+++ server/internal/background/activities.go
@@ -77,7 +77,6 @@ type Activities struct {
 	signalAssistantCoordinator      *activities.SignalAssistantCoordinator
 	signalAssistantThread           *activities.SignalAssistantThread
 	processWorkOSOrganizationEvents *activities.ProcessWorkOSOrganizationEvents
-	processWorkOSMembershipEvents   *activities.ProcessWorkOSMembershipEvents
 	processWorkOSGlobalRoleEvents   *activities.ProcessWorkOSGlobalRoleEvents
 	cancelAssistantsSubscription    *activities.CancelAssistantsSubscription
 }
@@ -119,11 +118,9 @@ func NewActivities(
 	// configured. Local dev without a key passes nil; the wrapper method below
 	// returns a clear error if the activity is invoked unconfigured.
 	var processWorkOSOrgEvents *activities.ProcessWorkOSOrganizationEvents
-	var processWorkOSMembershipEvents *activities.ProcessWorkOSMembershipEvents
 	var processWorkOSGlobalRoleEvents *activities.ProcessWorkOSGlobalRoleEvents
 	if workosEventsClient != nil {
 		processWorkOSOrgEvents = activities.NewProcessWorkOSOrganizationEvents(logger, db, workosEventsClient)
-		processWorkOSMembershipEvents = activities.NewProcessWorkOSMembershipEvents(logger, db, workosEventsClient)
 		processWorkOSGlobalRoleEvents = activities.NewProcessWorkOSGlobalRoleEvents(logger, db, workosEventsClient)
 	}
 
@@ -166,7 +163,6 @@ func NewActivities(
 		signalAssistantCoordinator:      activities.NewSignalAssistantCoordinator(&AssistantWorkflowSignaler{TemporalEnv: temporalEnv}),
 		signalAssistantThread:           activities.NewSignalAssistantThread(&AssistantWorkflowSignaler{TemporalEnv: temporalEnv}),
 		processWorkOSOrganizationEvents: processWorkOSOrgEvents,
-		processWorkOSMembershipEvents:   processWorkOSMembershipEvents,
 		processWorkOSGlobalRoleEvents:   processWorkOSGlobalRoleEvents,
 		cancelAssistantsSubscription:    activities.NewCancelAssistantsSubscription(logger, billingRepo),
 	}
@@ -177,13 +173,6 @@ func (a *Activities) ProcessWorkOSOrganizationEvents(ctx context.Context, params
 		return nil, fmt.Errorf("WorkOS events client is not configured")
 	}
 	return a.processWorkOSOrganizationEvents.Do(ctx, params)
-}
-
-func (a *Activities) ProcessWorkOSMembershipEvents(ctx context.Context, params activities.ProcessWorkOSMembershipEventsParams) (*activities.ProcessWorkOSMembershipEventsResult, error) {
-	if a.processWorkOSMembershipEvents == nil {
-		return nil, fmt.Errorf("WorkOS events client is not configured")
-	}
-	return a.processWorkOSMembershipEvents.Do(ctx, params)
 }
 
 func (a *Activities) ProcessWorkOSGlobalRoleEvents(ctx context.Context, params activities.ProcessWorkOSGlobalRoleEventsParams) (*activities.ProcessWorkOSGlobalRoleEventsResult, error) {
